CMake: properly check for variables
Checking them for being defined would get the user intention wrong if they are
passed as -DDISABLE_STATIC=Off as that would still count as being defined. Just
check for the variable, which has the desired effect. An unset variable is
always false. Since these flags are options now the variables are always set,
so one cannot disable this features anymore otherwise.