Raise an error on duplicate local variable name
Prevent local variable shadowing, deny redeclaring local variables
with the same name in the same scope.
This includes function arguments, local variables and foreach
iterators.
This is a breaking change - existing scripts can fail now.
But making the language stricter is a good thing as it helps
to avoid errors.

Example code (was compiling earlier, but not now)

local function foo(player) {
  local player = get_local_player()
  foreach (id, player in enemies) {
    player.respawn()
  }
}

or simply

local a = 10
local x = "foo"
local a = "bar"
local nextA = a-1

It is better to find such (potential) errors earlier, in compile time
instead of runtime.