Check for duplicate keys in table/class initialization
Original compiler behavior for table/class initializer is
to generate code that overwrites existing slots with the
same key.
Most probably this is an error:

  local config = {
    timeout = 100
    foo = 234
    bar = "some_string"
    timeout = 0
  }

  class Foo {
    function onUpdate() {
      doOneThing()
    }

    ...

    function onUpdate() {
      doCompletelyDifferentThing()
    }
  }

To prevent possible errors, make the compiler stricter and raise
a compile error when duplicate keys are detected.

When the key is a constant string the implementation is obvious,
however, when [] syntax is used, the key can be an arbitrary
expression. In this case in general we cannot validate everything
in compile time, but for the case of single string literal or
integer check if a single LOAD instruction was generated, thus
being able to detect collisions of the keys like this:

  local table = {
    ["timeout"] = 123
    ["timeout"] = 456
    [5] = "foo"
    [5] = "bar"
  }