Allow naming functions in expressions
And for unnamed functions generate surrogate name from source
file name and line which is much more useful than empty string.

Allow not only

  local a = function() {
    ::print("function name = "+::callee().getfuncinfos().name+"\n")
   }

or for example

  bar(function() {})

but support specifying optional function

  name local a = function foo() {
    ::print("function name = "+::callee().getfuncinfos().name+"\n")
  }

or

  bar(function foo() {})

When called a() in first case it will print surrogate function name generated from
source file and line, and in second it will use specified name (foo).

This doesn't interfere with local variables.