Introduce immutable objects
Allow to 'freeze' objects which will raise a runtime error on
trying to modify them.

Immutability flags is applied to SQObject, which is a handle
referencing an actual container, just like const reference in C++.
So the object possibly can be mutated by the external code (again
like in C++)

The immutability is propagated to nested fields during get operation,
so if table t={w={x=123}} is immutable, t.w is also immutable.

A new _flags field was added added to SQObject.
When SQInteger is 64-bit (which in case of Dagor engine is true
both on 32 and 64 bit platform) it doesn't affect size of SQObject
struct.
On 32 bit platform with 32-bit SQInteger it currently increases
struct size from 8 to 12 bytes. If needed _flags may be packed
together with _type field (since together they need 28+1 bits),
or alternatively NaN tagging may be used.

Squirrel API added:

  freeze(object) - make a immutable object handle
  getobjflags(object) - return object flags

Also SQOBJ_FLAG_IMMUTABLE was added to const table