Add sqstd_formatcallstackstring() function
It is often needed to receive stack info in a readable form
but before there was only sqstd_printcallstack() function
which prints formatted info by calling error printing function.
This may be inconvenient when we don't know what error printing
function does and have to temporary replace it with our custom one
that accumulates intermediate strings into one buffer.

sqstd_formatcallstackstring() builds a string and pushes it
to the VM stack.