Fix top-level function missing in call stack dump
Printing call stack may be performed not only by call of the
native function, but possibly also by such tools as debugger.
In this case the top-level function in call stack is an actual
script function being executed and should not be skipped.
So start stack traversal from level 0 and check if the top
is a script or a native function and include the script one
into output.