Forbid implicit global functions and classes by default
Declaring functions and classes in 'this'/root considered harmful.

In future we want to make function and class declarations local, but
since this is a breaking change, first let's forbid implicit global
functions and classes in compiler and when all projects are proved
to be working, the change in language can be made.

For now, if old behavior is needed, use #allow-func-decl-sugar and

Old behavior:

    function foo() {}

    // this is equivalent to

    this["foo"] <- function foo() {}

Behavior with this commit:

    function foo() {} // compilation error

Future planned behavior:

    function foo() {}

    // equivalent to

    let function foo() {}

    // this also mean 'let function name' and 'local function name'
    // support may be also removed