Remove class metamethods (_inherited, _newmember)
They were added for unclear reason, for some specific use case that
seems to be possible to implement in script itself without
complicating the compiler, VM and base library.
They were never used in Gaijin codebase (that is more than 17 years)
so let's keep the language core as simple as possible to ease
code maintenance.