Drop support for per-closure root
It was added in Squirrel version 3.1 beta 1.
However even after making access to `this` and root table explicit
there are problems with this feature in that the closure still has
too many contexts:
  * Captured free variables
  * `this`
  * Root table
Not only this is too much (so that support for root table access is
disabled in all new projects) and using free variables is more than
enough.
It is also less maintainable and more error-prone, because `this`
may be overridden dynamically in runtime using bindenv() and the
per-closure root before this commit could be replaced using
setroot() function.
The fact that the value of `this` or the result of :: operator may
be redefined by arbitrary script code at any moment is totally
not obvious for those who read the code.
It also was never used in Gaijin codebase except a few places
where it actually was unnecessary.

So, according Zen of Python principles
  Explicit is better than implicit.
  Simple is better than complex.
  There should be one-- and preferably only one --obvious way to do it.
remove support for per-closure root. Free variables should be used
for closure context from now on.

Change-Id: I219eb2766749c1ea656e3bd05f90dcdfab6a22a6