Add support for _SQ64 on 32-bit architectures.
This should be automatically detected so defining _SQ64 is now safe on
32-bit architectures. If it doesn't auto-detect (because you're using a
compiler that doesn't have a _LP64-like define) the build will error
(using a static assert in compilers that support it and a negative array
size in compilers that don't) and you can fix this by defining _SQ64ON32
or _NOSQ64ON32 (depending on which direction it misdetected in).

This means that a piece of software that is distributed on both 32 and
64 bit architectures can use a single set of compiled scripts, assuming
the size of SQFloat and SQChar also match.